fix: name memo category section 상태 lifting
