+++ snapsplit/src/features/trip/[tripId]/budget/expense/_components/expense-form/CategorySection.tsx
@@ -3,13 +3,18 @@
 import CategoryItem from './category/CategoryItem';
 import { EXPENSE_CATEGORY } from '@/shared/constants/expense';
 
-export default function CategorySection() {
+type Props = {
+  category: string;
+  setCategory: (category: string) => void;
+};
+
+export default function CategorySection({ category, setCategory }: Props) {
   return (
     <div className="flex flex-col items-start w-full gap-3">
       <div className="text-body-3">지출 카테고리</div>
       <div className="flex items-center gap-2 w-full">
-        {EXPENSE_CATEGORY.map((category) => (
-          <CategoryItem key={category.name} name={category.name} />
+        {EXPENSE_CATEGORY.map((c) => (
+          <CategoryItem key={c.name} name={c.name} category={category} setCategory={setCategory} />
         ))}
       </div>
     </div>
